+++ src/api/SSDPDB_ObjC.h
@@ -66,8 +66,8 @@
 -(int)searchForMediaServer;
 -(int)notifySSDPAlive;
 -(int)notifySSDPByeBye;
--(int)addObserver:(SSDPDB_ObjC_Observer*)obs;
--(int)removeObserver:(SSDPDB_ObjC_Observer*)obs;
+-(NSUInteger)addObserver:(SSDPDB_ObjC_Observer*)obs;
+-(NSUInteger)removeObserver:(SSDPDB_ObjC_Observer*)obs;
 -(void)SSDPDBUpdate;
 -(void)setUserAgentProduct:(NSString*)product andOS:(NSString*)os;
